Read Tp

This helper function reads the number of choice occasions Tp from a choice_identifiers object.

Usage

read_Tp(choice_identifiers)

Arguments

choice_identifiers

[choice_identifiers]
The choice_identifiers object that defines the choice occasions.

Value

An integer vector of length N, where N is the number of deciders.